Test de ver categoria de la persona juridica
