+++ tests/00-nombresyapellidos-personas.ts
@@ -1,22 +1,22 @@
 // Archivo para colocar los nombres de las personas a crear, necesarias para poder correr todos los tests
 
 // Persona Fisica Mayor de edad. Test = 05-registrar-persona-fisica.test.ts
-export const nombrePersonaFisica:string = '';
+export const nombrePersonaFisica:string = ''; // (Femenino)
 export const apellidoPersonaFisica:string = '';
 
 // Persona Juridica y Relacionado. Test = 06-registrar-persona-juridica.test.ts
 export const nombreJuridica:string = '';
 
-export const nombreRelacionadoJuridica:string = '';
-export const apellidoRelacionadoJuridica:string = '';
+export const nombreRelacionadoJuridica:string = ''; // (Masculino)
+export const apellidoRelacionadoJuridica:string = ''; 
 
 // Persona Fisica Menor de edad. Test = 07-registrar-persona-fisica-menor-edad.test.ts
-export const nombrePersonaMenorEdad:string = '';
-export const apellidoPersonaMenorEdad:string = '';
+export const nombrePersonaMenorEdad:string = ''; // (Masculino)
+export const apellidoPersonaMenorEdad:string = ''; 
 
 // Persona Fisica Casada. Test = 09-registrar-persona-fisica-casada.test.ts
-export const nombrePersonaFisicaCasada:string = '';
+export const nombrePersonaFisicaCasada:string = ''; // (Femenino)
 export const apellidoPersonaFisicaCasada:string = '';
 
-export const nombrePersonaFisicaConyuge:string = '';
+export const nombrePersonaFisicaConyuge:string = ''; // (Masculino)
 export const apellidoPersonaFisicaConyuge:string = '';
