+++ tests/38-ahorros-por-nomina.spec.ts
@@ -1,4 +1,4 @@
-import { APIResponse, Browser, BrowserContext, chromium, Page, expect, Locator, test } from '@playwright/test';
+import { APIResponse, Browser, BrowserContext, chromium, expect, Page, Locator, test } from '@playwright/test';
 import { url_base, EscenariosPruebaCrearCuentas, ariaCerrar, selectBuscar } from './utils/dataTests';
 
 // Variables globales
@@ -24,21 +24,21 @@ let apellidoFirmante: string | null;
 
 // Pruebas
 
-test.describe('Certificados - Inversion Pagaderas - Pruebas con los diferentes parametros', async () => {
+test.describe.serial('Crear Cuenta de Ahorros - Ahorros por Nomina - Pruebas con los diferentes parametros', async () => {
     for (const escenario of EscenariosPruebaCrearCuentas) {
         test.describe(`Test cuando el escenario es: ${Object.values(escenario).toString()}`, () => {
-            test.beforeAll(async () => { // Antes de las pruebas
-                // Crear el browser
+            test.beforeAll(async () => { // Antes de todas las pruebas
+                /* Crear el browser, con la propiedad headless */
                 browser = await chromium.launch({
                     headless: false
                 });
         
-                // Crear el context
+                /* Crear un context con el storageState donde esta guardado el token de la sesion */
                 context = await browser.newContext({
                     storageState: 'state.json'
                 });
         
-                // Crear una nueva page
+                /* Crear una nueva page usando el context */
                 page = await context.newPage();
 
                 // Eventos para la request relation
@@ -50,25 +50,25 @@ test.describe('Certificados - Inversion Pagaderas - Pruebas con los diferentes p
                     const body = await response.json();
                     // Condicion para cambiar los parametros del body
                     if (Object.keys(body?.data[33]).length > 1) {
-                        // Remplazar el body con la response con los datos de los escenarios
+                        // Reemplazar el body con la response con los datos del escenario
                         body.data[33] = Object.assign(body.data[33], escenario);
                         route.fulfill({
                             response,
-                            body: JSON.stringify(body),
+                            body: JSON.stringify(body)
                         })
                     } else {
                         route.continue();
                     };
                 });
         
-                // Ingresar a la pagina
+                /* Ingresar a la pagina */
                 await page.goto(`${url_base}`);
 
                 // Boton de Crear Nueva Cuenta
                 botonNuevaCuenta = page.getByRole('button', {name: 'plus Nueva Cuenta'});
 
                 // Cedula, nombre y apellido de la persona almacenada en el state
-                cedula = await page.evaluate(() => window.localStorage.getItem('cedula'));
+                cedula = await page.evaluate(() => window.localStorage.getItem('cedulaPersona'));
                 nombre = await page.evaluate(() => window.localStorage.getItem('nombrePersona'));
                 apellido = await page.evaluate(() => window.localStorage.getItem('apellidoPersona'));
 
@@ -86,24 +86,24 @@ test.describe('Certificados - Inversion Pagaderas - Pruebas con los diferentes p
                 await botonContinuar.click();
             };
         
-            test('Ir a la opcion de Apertura de cuentas de Certificados', async () => {
+            test('Ir a la opcion de Apertura de cuentas de Ahorros', async () => {
                 // Boton de Captaciones
                 await page.getByRole('menuitem', {name: 'CAPTACIONES'}).click();
         
                 // Boton de Apertura de cuentas
                 await page.getByRole('menuitem', {name: 'APERTURA DE CUENTAS'}).click();
         
-                // Boton de Certificados
-                await page.getByRole('menuitem', {name: 'Certificados'}).click();
+                // Boton de Ahorros
+                await page.getByRole('menuitem', {name: 'Ahorros'}).click();
         
                 // La url debe de cambiar
-                await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados`);
+                await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-2/ahorros`);
         
-                // El titulo de Certificados debe estar visible
-                await expect(page.locator('h1').filter({hasText: 'CERTIFICADOS'})).toBeVisible();
+                // El titulo de ahorros debe estar visible
+                await expect(page.locator('h1').filter({hasText: 'AHORROS'})).toBeVisible();
             });
         
-            test('Seleccionar el Certificado Inversion Pagaderas', async () => {
+            test('Seleccionar un tipo de captaciones', async () => {
                 // El titulo de tipo de captaciones debe estar visible
                 await expect(page.locator('h1').filter({hasText: 'TIPO DE CAPTACIONES'})).toBeVisible();
         
@@ -113,16 +113,16 @@ test.describe('Certificados - Inversion Pagaderas - Pruebas con los diferentes p
                 // Click al boton
                 await botonCaptaciones.click();
 
-                // Click a la opcion de Inversion Pagaderas
-                const opcionInversionPagaderas = page.locator('text=INVERSION PAGADERAS');
-                await expect(opcionInversionPagaderas).toBeVisible();
-                await opcionInversionPagaderas.click();
+                // Click a la opcion de Ahorros por Nomina
+                const opcionAhorrosNomina = page.locator('text=AHORROS POR NOMINA');
+                await expect(opcionAhorrosNomina).toBeVisible();
+                await opcionAhorrosNomina.click();
         
                 // La URL debe de cambiar al elegir el tipo de captacion
-                await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/23`);
+                await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-2/ahorros/19`);
 
-                // El tipo de captacion de Inversion Pagaderas debe estar visible
-                await expect(page.locator('#form').getByTitle('INVERSION PAGADERAS')).toBeVisible();
+                // El tipo de captacion de ahorros por nomina debe estar visible
+                await expect(page.locator('#form').getByTitle('AHORROS POR NOMINA')).toBeVisible();
             });
 
             if (escenario.ID_OPERACION === '') {
@@ -157,119 +157,58 @@ test.describe('Certificados - Inversion Pagaderas - Pruebas con los diferentes p
                 });
             } else if (escenario.ID_OPERACION === 30) {
                 // Tests si el ID_OPERACION es 30
-                test('Boton de Nueva Cuenta', async () => {
+                test('Click al boton de Nueva Cuenta', async () => {
+                    // Boton de Nueva Cuenta
                     await expect(botonNuevaCuenta).toBeVisible();
                     await botonNuevaCuenta.click();
+            
+                    // La URL debe de cambiar
+                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-2/ahorros/19/create?step=1`);
+            
+                    // El titulo de Registrar Cuenta debe estar visible
+                    await expect(page.locator('text=CREAR CUENTA DE AHORROS')).toBeVisible();
                 });
-
-                test('Crear una Nueva Cuenta de Certificado - Paso 1 - Datos Generales', async () => {
-                    // El titulo debe estar visible
-                    await expect(page.locator('h1').filter({hasText: 'CREAR CUENTA DE CERTIFICADOS'})).toBeVisible();
             
-                    // La URL debe cambiar
-                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/23/create?step=1`);
-            
-                    // La cuenta debe ser de inversion pagaderas
-                    await expect(page.locator('text=INVERSION PAGADERAS').first()).toBeVisible();
-            
+                test('Llenar los campos del primer paso del registro de cuenta de ahorros - orden de pago', async () => {
                     // Titular
-                    const campoTitular = page.locator(`${selectBuscar}`).first();
-            
+                    const campoTitular = page.locator(`${selectBuscar}`);
+                    
+                    // Buscar un socio
                     await campoTitular?.fill(`${cedula}`);
                     // Seleccionar la opcion que aparece
-                    await page.locator(`text=${cedula}`).click(); 
+                    await page.locator(`text=${cedula}`).click();
             
-                    // Cambiar la descripcion de la cuenta
-                    const descripcionCuenta = page.getByPlaceholder('Descripción o alias de la cuenta, ejemplo: Cuenta para vacaciones.');
-                    // Viene con una descripcion por defecto, borrar dicha descripcion
-                    await descripcionCuenta.clear();
-                    // Nueva descripcion de la cuenta
-                    await descripcionCuenta.fill('Cuenta de certificado inversion pagadera');
+                    // El tipo de captacion debe ser Ahorros
+                    await expect(page.locator('text=AHORROS POR NOMINA').first()).toBeVisible();
             
-                    // La categoria del socio debe ser socio ahorrante
-                    await expect(page.locator('text=SOCIO AHORRANTE')).toBeVisible();
+                    // Subir la imagen de la firma
+                    const subirFirmaPromesa = page.waitForEvent('filechooser'); // Esperar por el evento de filechooser
+                    await page.getByText('Cargar ').click(); 
+                    const subirFirma = await subirFirmaPromesa; // Guardar el evento del filechooser en una constante
+                    await subirFirma.setFiles(`${firma}`); // setFiles para elegir un archivo
             
-                    // Plazo
-                    await page.getByPlaceholder('PLAZO').fill('24');
-            
-                    // El plazo debe ser mensual, que es el que viene por defecto
-                    await expect(page.locator('text=MENSUAL')).toBeVisible();
-            
-                    // Ingresar un monto
-                    const campoMonto = page.getByPlaceholder('MONTO');
-                    await campoMonto.clear();
-                    await campoMonto.fill('50');
-            
-                    // Desmarcar el via de cobro, debito a cuenta
-                    const casillaDebitoCuenta = page.getByLabel('Débito a cuenta(s)');
-                    await casillaDebitoCuenta.click();
-            
-                    // Debe salir una advertencia
-                    await expect(page.locator('text=Via Cobro es requerido.')).toBeVisible();
-            
-                    // Marcar la casilla de debito a cuenta
-                    await casillaDebitoCuenta.click();
-            
-                    // Ingresar la tasa
-                    await page.locator('#INVERSION\\ PAGADERAS_TASA').fill('5');
-            
-                    // Click al boton de cargar autorizacion
-                    await expect(page.getByRole('button', {name: 'Cargar Autorización'})).toBeVisible();
-
-                    // El boton de subir la firma no debe estar visible
-                    await expect(page.getByRole('button', {name: 'upload Cargar'}).getByRole('button', {name: 'upload Cargar', exact: true}).filter({hasText: 'Cargar'})).not.toBeVisible();
-            
-                    // El titulo de origen de inversion debe estar visible
-                    await expect(page.locator('h1').filter({hasText: 'ORIGEN DE INVERSIÓN'})).toBeVisible();
-            
-                    // Buscar una cuenta de la persona
-                    const campoBuscarCuenta = page.locator(`${selectBuscar}`).last();
-                    await campoBuscarCuenta.click();
-                    // Elegir la cuenta de ahorros
-                    await page.locator('text=AHORROS NORMALES').click();
-            
-                    // Boton Agregar la cuenta
-                    const botonAgregar = page.getByRole('button', {name: 'plus Agregar'});
-                    await expect(botonAgregar).toBeVisible();
-                    // Click al boton 
-                    await botonAgregar.click();
-            
-                    // El monto por defecto es el monto de apertura, clickear en otro lugar para que se guarde el monto
-                    await page.locator('text=TOTALES').click();
-            
-                    // Click al boton de Continuar
-                    const botonContinuar = page.getByRole('button', {name: 'Continuar'});
-                    // Esperar que se abra una nueva pestaña con el reporte de la nota de debito
-                    const [newPage] = await Promise.all([
-                        context.waitForEvent('page'),
-                        // Click al boton de Aceptar
-                        await expect(botonContinuar).toBeVisible(),
-                        await botonContinuar.click()
-                    ]);
-
-                    // Cerrar la pagina con el reporte de la nota de debito
-                    await newPage.close();
+                    // Click al boton de continuar
+                    Continuar();
                 });
             
-                test('Crear una Nueva Cuenta de Certificado - Paso 2 - Contacto de Firmante', async () => {            
-                    // La URL debe cambiar
-                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/23/create?step=2`);
+                test('Contacto de Firmante o Persona', async () => {            
+                    // La URL debe de cambiar
+                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-2/ahorros/19/create?step=2`);
             
                     // El titulo de firmantes debe estar visible
                     await expect(page.locator('h1').filter({hasText: 'FIRMANTES'})).toBeVisible();
             
-                    // Cambiar a la pestaña de Personas o Contactos
-                    const seccionPersonaContactos = page.locator('text=Personas o Contactos');
-                    await seccionPersonaContactos.click();
-            
-                    // Titulo de la seccion debe estar visible
-                    await expect(page.locator('h1').filter({hasText: 'CONTACTOS CON LA PERSONAS O EMPRESA'})).toBeVisible();
-            
-                    // Regresar a la seccion de firmantes
-                    await page.getByRole('tab').filter({hasText: 'Firmantes'}).click();
-            
-                    // Cerrar los mensajes que aparecen
+                    // Cerrar uno de los mensajes que aparecen
                     await page.locator(`${ariaCerrar}`).first().click();
+
+                    // El titular debe estar en la tabla de los firmantes
+                    await expect(page.getByRole('cell', {name: `${nombre} ${apellido}`})).toBeVisible();
+
+                    // Se debe mostrar la firma del titular por defecto
+                    await expect(page.locator('text=TITULAR')).toBeVisible();
+            
+                    // El tipo de firma requerida debe estar visible
+                    await expect(page.locator('text=(Y) FIRMA REQUERIDA')).toBeVisible();
             
                     // Boton de Agregar Firmantes debe estar visible
                     const botonAgregarFirmantes = page.locator('text=Agregar Firmante');
@@ -285,6 +224,7 @@ test.describe('Certificados - Inversion Pagaderas - Pruebas con los diferentes p
                     await buscador.click();
                     await buscador.fill(`${cedulaFirmante}`);
                     // Seleccionar el socio
+                    await expect(page.locator(`text=${nombreFirmante} ${apellidoFirmante}`)).toBeVisible();
                     await page.locator(`text=${nombreFirmante} ${apellidoFirmante}`).click();
             
                     // Debe salir otro modal para llenar la informacion de la firmante
@@ -342,42 +282,43 @@ test.describe('Certificados - Inversion Pagaderas - Pruebas con los diferentes p
                     Continuar();
                 });
             
-                test('Crear una Nueva Cuenta de Certificado - Paso 3 - Metodo de Interes', async () => {
-                    // La URL debe cambiar
-                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/23/create?step=3`);
-            
-                    // El titulo principal debe estar visible
+                test('Metodo de intereses', async () => {
+                    // La URL debe de cambiar
+                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-2/ahorros/19/create?step=3`);
+                    
+                    // El titulo debe estar visible
                     await expect(page.locator('h1').filter({hasText: 'FORMA PAGO DE INTERESES O EXCEDENTES'})).toBeVisible();
 
                     // Debe mostrarse la cuenta que se esta creando, y el titular
                     await expect(page.getByRole('cell', {name: `${nombre} ${apellido}`})).toBeVisible();
                 });
             
-                test('Finalizar con la Creacion de Cuenta de Certificado', async () => {
+                test('Finalizar con el registro de cuenta de ahorro', async () => {
                     // Esperar que el mensaje de que los contratos se hayan generado se muestre
                     await expect(page.locator('text=Contratos Generados Exitosamente.')).toBeVisible();
 
                     // Boton de Finalizar
-                    const botonFinalizar = page.locator('button:has-text("Finalizar")');
-                    // Esperar que se abran dos pestañas
-                    const [newPage, newPage2] = await Promise.all([
-                        context.waitForEvent('page'),
+                    const botonFinalizar = page.getByRole('button', {name: 'Finalizar'});
+                    // Esperar que se abra una nueva pestaña
+                    const [newPage] = await Promise.all([
                         context.waitForEvent('page'),
                         // Click al boton de Finalizar
                         await expect(botonFinalizar).toBeVisible(),
                         await botonFinalizar.click()
                     ]);
                   
-                    // Cerrar las dos paginas abiertas
+                    // La pagina abierta con la solicitud se cierra
                     await newPage.close();
-                    await newPage2.close();
-
-                    // Debe regresar a la pagina de los certificados
-                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/23`);
+                    
+                    // Debe de regresar a la pagina las cuentas de ahorros
+                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-2/ahorros/19`);
+            
+                    // El titulo de Ahorros debe estar visible
+                    await expect(page.locator('h1').filter({hasText: 'AHORROS'})).toBeVisible();
                 });
             };
-            
-            test.afterAll(async () => { // Despues de las pruebas
+        
+            test.afterAll(async () => { // Despues de todas las pruebas
                 // Cerrar la page
                 await page.close();
