+++ tests/42-certificados-financieros-reinvertidas.spec.ts
@@ -1,4 +1,4 @@
-import { APIResponse, Browser, BrowserContext, chromium, expect, Page, Locator, test } from '@playwright/test';
+import { APIResponse, Browser, BrowserContext, chromium, Page, expect, Locator, test } from '@playwright/test';
 import { url_base, EscenariosPruebaCrearCuentas, ariaCerrar, selectBuscar } from './utils/dataTests';
 
 // Variables globales
@@ -24,21 +24,21 @@ let apellidoFirmante: string | null;
 
 // Pruebas
 
-test.describe.serial('Crear Cuenta de Ahorros - Orden de Pago - Pruebas con los diferentes parametros', async () => {
+test.describe.serial('Certificados - Financieros Reinvertidas - Pruebas con los diferentes parametros', async () => {
     for (const escenario of EscenariosPruebaCrearCuentas) {
         test.describe(`Test cuando el escenario es: ${Object.values(escenario).toString()}`, () => {
-            test.beforeAll(async () => { // Antes de todas las pruebas
-                /* Crear el browser, con la propiedad headless */
+            test.beforeAll(async () => { // Antes de las pruebas
+                // Crear el browser
                 browser = await chromium.launch({
-                    headless: false
+                    headless: false,
                 });
         
-                /* Crear un context con el storageState donde esta guardado el token de la sesion */
+                // Crear el context
                 context = await browser.newContext({
                     storageState: 'state.json'
                 });
         
-                /* Crear una nueva page usando el context */
+                // Crear una nueva page
                 page = await context.newPage();
 
                 // Eventos para la request relation
@@ -50,18 +50,18 @@ test.describe.serial('Crear Cuenta de Ahorros - Orden de Pago - Pruebas con los 
                     const body = await response.json();
                     // Condicion para cambiar los parametros del body
                     if (Object.keys(body?.data[33]).length > 1) {
-                        // Reemplazar el body con la response con los datos del escenario
+                        // Remplazar el body con la response con los datos de los escenarios
                         body.data[33] = Object.assign(body.data[33], escenario);
                         route.fulfill({
                             response,
-                            body: JSON.stringify(body)
+                            body: JSON.stringify(body),
                         })
                     } else {
                         route.continue();
                     };
                 });
         
-                /* Ingresar a la pagina */
+                // Ingresar a la pagina
                 await page.goto(`${url_base}`);
 
                 // Boton de Crear Nueva Cuenta
@@ -86,24 +86,24 @@ test.describe.serial('Crear Cuenta de Ahorros - Orden de Pago - Pruebas con los 
                 await botonContinuar.click();
             };
         
-            test('Ir a la opcion de Apertura de cuentas de Ahorros', async () => {
+            test('Ir a la opcion de Apertura de cuentas de Certificados', async () => {
                 // Boton de Captaciones
                 await page.getByRole('menuitem', {name: 'CAPTACIONES'}).click();
         
                 // Boton de Apertura de cuentas
                 await page.getByRole('menuitem', {name: 'APERTURA DE CUENTAS'}).click();
         
-                // Boton de Ahorros
-                await page.getByRole('menuitem', {name: 'Ahorros'}).click();
+                // Boton de Certificados
+                await page.getByRole('menuitem', {name: 'Certificados', exact: true}).click();
         
                 // La url debe de cambiar
-                await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-2/ahorros`);
+                await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados`);
         
-                // El titulo de ahorros debe estar visible
-                await expect(page.locator('h1').filter({hasText: 'AHORROS'})).toBeVisible();
+                // El titulo de Certificadoss debe estar visible
+                await expect(page.locator('h1').filter({hasText: 'CERTIFICADOS'})).toBeVisible();
             });
         
-            test('Seleccionar un tipo de captaciones', async () => {
+            test('Seleccionar el Certificado Financieros Reinvertidas', async () => {
                 // El titulo de tipo de captaciones debe estar visible
                 await expect(page.locator('h1').filter({hasText: 'TIPO DE CAPTACIONES'})).toBeVisible();
         
@@ -113,16 +113,16 @@ test.describe.serial('Crear Cuenta de Ahorros - Orden de Pago - Pruebas con los 
                 // Click al boton
                 await botonCaptaciones.click();
 
-                // Click a la opcion de Ahorros Normales
-                const opcionAhorrosNormales = page.locator('text=AHORROS NORMALES');
-                await expect(opcionAhorrosNormales).toBeVisible();
-                await opcionAhorrosNormales.click();
+                // Click a la opcion de Financieros Reinvertidas
+                const opcionFinancierosReinvertidas = page.locator('text=FINANCIEROS REINVERTIDAS');
+                await expect(opcionFinancierosReinvertidas).toBeVisible();
+                await opcionFinancierosReinvertidas.click();
         
                 // La URL debe de cambiar al elegir el tipo de captacion
-                await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-2/ahorros/16`);
+                await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/10`);
 
-                // El tipo de captacion de ahorros normales debe estar visible
-                await expect(page.locator('#form').getByTitle('AHORROS NORMALES')).toBeVisible();
+                // El tipo de captacion de Financieros Reinvertidas debe estar visible
+                await expect(page.locator('#form').getByTitle('FINANCIEROS REINVERTIDAS')).toBeVisible();
             });
 
             if (escenario.ID_OPERACION === '') {
@@ -157,49 +157,136 @@ test.describe.serial('Crear Cuenta de Ahorros - Orden de Pago - Pruebas con los 
                 });
             } else if (escenario.ID_OPERACION === 30) {
                 // Tests si el ID_OPERACION es 30
-                test('Click al boton de Nueva Cuenta', async () => {
+                test('Boton de Nueva Cuenta', async () => {
                     // Boton de Nueva Cuenta
                     await expect(botonNuevaCuenta).toBeVisible();
                     await botonNuevaCuenta.click();
-            
-                    // La URL debe de cambiar
-                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-2/ahorros/17/create?step=1`);
-            
-                    // El titulo de Registrar Cuenta debe estar visible
-                    await expect(page.locator('text=CREAR CUENTA DE AHORROS')).toBeVisible();
                 });
+
+                test('Crear una Nueva Cuenta de Certificado - Paso 1 - Datos Generales', async () => {
+                    // El titulo debe estar visible
+                    await expect(page.locator('h1').filter({hasText: 'CREAR CUENTA DE CERTIFICADOS'})).toBeVisible();
             
-                test('Llenar los campos del primer paso del registro de cuenta de ahorros - orden de pago', async () => {
+                    // La URL debe cambiar
+                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/10/create?step=1`);
+            
+                    // La cuenta debe ser de financieros reinvertidas
+                    await expect(page.locator('text=FINANCIEROS REINVERTIDAS').first()).toBeVisible();
+            
                     // Titular
-                    const campoTitular = page.locator(`${selectBuscar}`);
-                    
-                    // Buscar un socio
+                    const campoTitular = page.locator(`${selectBuscar}`).first();
+            
                     await campoTitular?.fill(`${cedula}`);
                     // Seleccionar la opcion que aparece
-                    await page.locator(`text=${cedula}`).click();
+                    await page.locator(`text=${cedula}`).click(); 
             
-                    // El tipo de captacion debe ser Ahorros
-                    await expect(page.locator('text=ORDEN DE PAGO').first()).toBeVisible();
+                    // Cambiar la descripcion de la cuenta
+                    const descripcionCuenta = page.getByPlaceholder('Descripción o alias de la cuenta, ejemplo: Cuenta para vacaciones.');
+                    // Viene con una descripcion por defecto, borrar dicha descripcion
+                    await descripcionCuenta.clear();
+                    // Nueva descripcion de la cuenta
+                    await descripcionCuenta.fill('Cuenta de certificado financiero reinvertida');
             
-                    // Subir la imagen de la firma
-                    const subirFirmaPromesa = page.waitForEvent('filechooser'); // Esperar por el evento de filechooser
-                    await page.getByText('Cargar ').click(); 
-                    const subirFirma = await subirFirmaPromesa; // Guardar el evento del filechooser en una constante
-                    await subirFirma.setFiles(`${firma}`); // setFiles para elegir un archivo
+                    // La categoria del socio debe ser socio ahorrante
+                    await expect(page.locator('text=SOCIO AHORRANTE')).toBeVisible();
             
-                    // Click al boton de continuar
-                    Continuar();
+                    // Plazo
+                    await page.getByPlaceholder('PLAZO').fill('36');
+            
+                    // El plazo debe ser mensual, que es el que viene por defecto
+                    await expect(page.locator('text=MENSUAL')).toBeVisible();
+            
+                    // Ver los rangos del monto de apertura
+                    await page.locator('[aria-label="eye"]').click();
+                    // Debe salir un modal
+                    const modalRangos = page.getByRole('heading', {name: 'Detalles de Rango'}).first();
+                    await expect(modalRangos).toBeVisible();
+            
+                    // Debe mostrar que el monto minimo es 1 peso dominicano
+                    await expect(page.getByRole('cell', {name: 'RD$ 1.00'}).nth(1)).toBeVisible();
+            
+                    // Click en Aceptar
+                    await page.getByRole('button', {name: 'check Aceptar'}).nth(1).click();
+            
+                    // El modal se debe cerrar
+                    await expect(modalRangos).not.toBeVisible();
+            
+                    // Ingresar un monto valido
+                    const campoMonto = page.getByPlaceholder('MONTO');
+                    await campoMonto.clear();
+                    await campoMonto.fill('50');
+            
+                    // Desmarcar el via de cobro, debito a cuenta
+                    const casillaDebitoCuenta = page.getByLabel('Débito a cuenta(s)');
+                    await casillaDebitoCuenta.click();
+            
+                    // Debe salir una advertencia
+                    await expect(page.locator('text=Via Cobro es requerido.')).toBeVisible();
+            
+                    // Marcar la casilla de debito a cuenta
+                    await casillaDebitoCuenta.click();
+            
+                    // Ingresar la tasa
+                    await page.locator('#FINANCIEROS\\ REINVERTIDAS_TASA').fill('8');
+            
+                    // Click al boton de cargar autorizacion
+                    await expect(page.getByRole('button', {name: 'Cargar Autorización'})).toBeVisible();
+
+                    // El boton de subir la firma no debe estar visible
+                    await expect(page.getByRole('button', {name: 'upload Cargar'}).getByRole('button', {name: 'upload Cargar', exact: true}).filter({hasText: 'Cargar'})).not.toBeVisible();
+            
+                    // El titulo de origen de inversion debe estar visible
+                    await expect(page.locator('h1').filter({hasText: 'ORIGEN DE INVERSIÓN'})).toBeVisible();
+            
+                    // Buscar una cuenta de la persona
+                    const campoBuscarCuenta = page.locator(`${selectBuscar}`).last();
+                    await campoBuscarCuenta.click();
+                    // Elegir la cuenta de ahorros
+                    await page.locator('text=AHORROS NORMALES').click();
+            
+                    // Boton Agregar la cuenta
+                    const botonAgregar = page.getByRole('button', {name: 'plus Agregar'});
+                    await expect(botonAgregar).toBeVisible();
+                    // Click al boton 
+                    await botonAgregar.click();
+            
+                    // El monto por defecto es el monto de apertura, clickear en otro lugar para que se guarde el monto
+                    await page.locator('text=TOTALES').click();
+            
+                    // Click al boton de Continuar
+                    const botonContinuar = page.getByRole('button', {name: 'Continuar'});
+                    // Esperar que se abra una nueva pestaña con el reporte de la nota de debito
+                    const [newPage] = await Promise.all([
+                        context.waitForEvent('page'),
+                        // Click al boton de Aceptar
+                        await expect(botonContinuar).toBeVisible(),
+                        await botonContinuar.click()
+                    ]);
+
+                    // Cerrar la pagina con el reporte de la nota de debito
+                    await newPage.close();
                 });
             
-                test('Contacto de Firmante o Persona', async () => {              
-                    // La URL debe de cambiar
-                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-2/ahorros/17/create?step=2`);
+                test('Crear una Nueva Cuenta de Certificado - Paso 2 - Contacto de Firmante', async () => {            
+                    // La URL debe cambiar
+                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/10/create?step=2`);
             
                     // El titulo de firmantes debe estar visible
                     await expect(page.locator('h1').filter({hasText: 'FIRMANTES'})).toBeVisible();
             
-                    // Cerrar uno de los mensajes que aparecen
+                    // Cambiar a la pestaña de Personas o Contactos
+                    const seccionPersonaContactos = page.locator('text=Personas o Contactos');
+                    await seccionPersonaContactos.click();
+            
+                    // Titulo de la seccion debe estar visible
+                    await expect(page.locator('h1').filter({hasText: 'CONTACTOS CON LA PERSONAS O EMPRESA'})).toBeVisible();
+            
+                    // Regresar a la seccion de firmantes
+                    await page.getByRole('tab').filter({hasText: 'Firmantes'}).click();
+            
+                    // Cerrar los mensajes que aparecen
                     await page.locator(`${ariaCerrar}`).first().click();
+                    await page.locator(`${ariaCerrar}`).last().click();
             
                     // Boton de Agregar Firmantes debe estar visible
                     const botonAgregarFirmantes = page.locator('text=Agregar Firmante');
@@ -215,7 +302,6 @@ test.describe.serial('Crear Cuenta de Ahorros - Orden de Pago - Pruebas con los 
                     await buscador.click();
                     await buscador.fill(`${cedulaFirmante}`);
                     // Seleccionar el socio
-                    await expect(page.locator(`text=${nombreFirmante} ${apellidoFirmante}`)).toBeVisible();
                     await page.locator(`text=${nombreFirmante} ${apellidoFirmante}`).click();
             
                     // Debe salir otro modal para llenar la informacion de la firmante
@@ -273,43 +359,42 @@ test.describe.serial('Crear Cuenta de Ahorros - Orden de Pago - Pruebas con los 
                     Continuar();
                 });
             
-                test('Metodo de intereses', async () => {
-                    // La URL debe de cambiar
-                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-2/ahorros/17/create?step=3`);
-                    
-                    // El titulo debe estar visible
+                test('Crear una Nueva Cuenta de Certificado - Paso 3 - Metodo de Interes', async () => {
+                    // La URL debe cambiar
+                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/10/create?step=3`);
+            
+                    // El titulo principal debe estar visible
                     await expect(page.locator('h1').filter({hasText: 'FORMA PAGO DE INTERESES O EXCEDENTES'})).toBeVisible();
 
                     // Debe mostrarse la cuenta que se esta creando, y el titular
                     await expect(page.getByRole('cell', {name: `${nombre} ${apellido}`})).toBeVisible();
                 });
             
-                test('Finalizar con el registro de cuenta de ahorro', async () => {
+                test('Finalizar con la Creacion de Cuenta de Certificado', async () => {
                     // Esperar que el mensaje de que los contratos se hayan generado se muestre
                     await expect(page.locator('text=Contratos Generados Exitosamente.')).toBeVisible();
-
+                    
                     // Boton de Finalizar
-                    const botonFinalizar = page.getByRole('button', {name: 'Finalizar'});
-                    // Esperar que se abra una nueva pestaña
-                    const [newPage] = await Promise.all([
+                    const botonFinalizar = page.locator('button:has-text("Finalizar")');
+                    // Esperar que se abran dos pestañas
+                    const [newPage, newPage2] = await Promise.all([
+                        context.waitForEvent('page'),
                         context.waitForEvent('page'),
                         // Click al boton de Finalizar
                         await expect(botonFinalizar).toBeVisible(),
                         await botonFinalizar.click()
                     ]);
                   
-                    // La pagina abierta con la solicitud se cierra
+                    // Cerrar las dos paginas abiertas
                     await newPage.close();
-                    
-                    // Debe de regresar a la pagina las cuentas de ahorros
-                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-2/ahorros/17`);
-            
-                    // El titulo de Ahorros debe estar visible
-                    await expect(page.locator('h1').filter({hasText: 'AHORROS'})).toBeVisible();
+                    await newPage2.close();
+
+                    // Debe regresar a la pagina de los certificados
+                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/10`);
                 });
             };
-        
-            test.afterAll(async () => { // Despues de todas las pruebas
+            
+            test.afterAll(async () => { // Despues de las pruebas
                 // Cerrar la page
                 await page.close();
 
@@ -319,4 +404,3 @@ test.describe.serial('Crear Cuenta de Ahorros - Orden de Pago - Pruebas con los 
         });
     };
 });
-
