+++ tests/43-certificados-inversion-pagaderas.spec.ts
@@ -24,13 +24,13 @@ let apellidoFirmante: string | null;
 
 // Pruebas
 
-test.describe.serial('Certificados - Financieros Reinvertidas - Pruebas con los diferentes parametros', async () => {
+test.describe('Certificados - Inversion Pagaderas - Pruebas con los diferentes parametros', async () => {
     for (const escenario of EscenariosPruebaCrearCuentas) {
         test.describe(`Test cuando el escenario es: ${Object.values(escenario).toString()}`, () => {
             test.beforeAll(async () => { // Antes de las pruebas
                 // Crear el browser
                 browser = await chromium.launch({
-                    headless: false,
+                    headless: false
                 });
         
                 // Crear el context
@@ -94,16 +94,16 @@ test.describe.serial('Certificados - Financieros Reinvertidas - Pruebas con los 
                 await page.getByRole('menuitem', {name: 'APERTURA DE CUENTAS'}).click();
         
                 // Boton de Certificados
-                await page.getByRole('menuitem', {name: 'Certificados'}).click();
+                await page.getByRole('menuitem', {name: 'Certificados', exact: true}).click();
         
                 // La url debe de cambiar
                 await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados`);
         
-                // El titulo de Certificadoss debe estar visible
+                // El titulo de Certificados debe estar visible
                 await expect(page.locator('h1').filter({hasText: 'CERTIFICADOS'})).toBeVisible();
             });
         
-            test('Seleccionar el Certificado Financieros Reinvertidas', async () => {
+            test('Seleccionar el Certificado Inversion Pagaderas', async () => {
                 // El titulo de tipo de captaciones debe estar visible
                 await expect(page.locator('h1').filter({hasText: 'TIPO DE CAPTACIONES'})).toBeVisible();
         
@@ -113,16 +113,16 @@ test.describe.serial('Certificados - Financieros Reinvertidas - Pruebas con los 
                 // Click al boton
                 await botonCaptaciones.click();
 
-                // Click a la opcion de Financieros Reinvertidas
-                const opcionFinancierosReinvertidas = page.locator('text=FINANCIEROS REINVERTIDAS');
-                await expect(opcionFinancierosReinvertidas).toBeVisible();
-                await opcionFinancierosReinvertidas.click();
+                // Click a la opcion de Inversion Pagaderas
+                const opcionInversionPagaderas = page.locator('text=INVERSION PAGADERAS');
+                await expect(opcionInversionPagaderas).toBeVisible();
+                await opcionInversionPagaderas.click();
         
                 // La URL debe de cambiar al elegir el tipo de captacion
-                await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/10`);
+                await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/23`);
 
-                // El tipo de captacion de Financieros Reinvertidas debe estar visible
-                await expect(page.locator('#form').getByTitle('FINANCIEROS REINVERTIDAS')).toBeVisible();
+                // El tipo de captacion de Inversion Pagaderas debe estar visible
+                await expect(page.locator('#form').getByTitle('INVERSION PAGADERAS')).toBeVisible();
             });
 
             if (escenario.ID_OPERACION === '') {
@@ -158,7 +158,6 @@ test.describe.serial('Certificados - Financieros Reinvertidas - Pruebas con los 
             } else if (escenario.ID_OPERACION === 30) {
                 // Tests si el ID_OPERACION es 30
                 test('Boton de Nueva Cuenta', async () => {
-                    // Boton de Nueva Cuenta
                     await expect(botonNuevaCuenta).toBeVisible();
                     await botonNuevaCuenta.click();
                 });
@@ -168,10 +167,10 @@ test.describe.serial('Certificados - Financieros Reinvertidas - Pruebas con los 
                     await expect(page.locator('h1').filter({hasText: 'CREAR CUENTA DE CERTIFICADOS'})).toBeVisible();
             
                     // La URL debe cambiar
-                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/10/create?step=1`);
+                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/23/create?step=1`);
             
-                    // La cuenta debe ser de financieros reinvertidas
-                    await expect(page.locator('text=FINANCIEROS REINVERTIDAS').first()).toBeVisible();
+                    // La cuenta debe ser de inversion pagaderas
+                    await expect(page.locator('text=INVERSION PAGADERAS').first()).toBeVisible();
             
                     // Titular
                     const campoTitular = page.locator(`${selectBuscar}`).first();
@@ -185,33 +184,18 @@ test.describe.serial('Certificados - Financieros Reinvertidas - Pruebas con los 
                     // Viene con una descripcion por defecto, borrar dicha descripcion
                     await descripcionCuenta.clear();
                     // Nueva descripcion de la cuenta
-                    await descripcionCuenta.fill('Cuenta de certificado financiero reinvertida');
+                    await descripcionCuenta.fill('Cuenta de certificado inversion pagadera');
             
                     // La categoria del socio debe ser socio ahorrante
                     await expect(page.locator('text=SOCIO AHORRANTE')).toBeVisible();
             
                     // Plazo
-                    await page.getByPlaceholder('PLAZO').fill('36');
+                    await page.getByPlaceholder('PLAZO').fill('24');
             
                     // El plazo debe ser mensual, que es el que viene por defecto
                     await expect(page.locator('text=MENSUAL')).toBeVisible();
             
-                    // Ver los rangos del monto de apertura
-                    await page.locator('[aria-label="eye"]').click();
-                    // Debe salir un modal
-                    const modalRangos = page.getByRole('heading', {name: 'Detalles de Rango'}).first();
-                    await expect(modalRangos).toBeVisible();
-            
-                    // Debe mostrar que el monto minimo es 1 peso dominicano
-                    await expect(page.getByRole('cell', {name: 'RD$ 1.00'}).nth(1)).toBeVisible();
-            
-                    // Click en Aceptar
-                    await page.getByRole('button', {name: 'check Aceptar'}).nth(1).click();
-            
-                    // El modal se debe cerrar
-                    await expect(modalRangos).not.toBeVisible();
-            
-                    // Ingresar un monto valido
+                    // Ingresar un monto
                     const campoMonto = page.getByPlaceholder('MONTO');
                     await campoMonto.clear();
                     await campoMonto.fill('50');
@@ -227,7 +211,7 @@ test.describe.serial('Certificados - Financieros Reinvertidas - Pruebas con los 
                     await casillaDebitoCuenta.click();
             
                     // Ingresar la tasa
-                    await page.locator('#FINANCIEROS\\ REINVERTIDAS_TASA').fill('8');
+                    await page.locator('#INVERSION\\ PAGADERAS_TASA').fill('5');
             
                     // Click al boton de cargar autorizacion
                     await expect(page.getByRole('button', {name: 'Cargar Autorización'})).toBeVisible();
@@ -269,7 +253,7 @@ test.describe.serial('Certificados - Financieros Reinvertidas - Pruebas con los 
             
                 test('Crear una Nueva Cuenta de Certificado - Paso 2 - Contacto de Firmante', async () => {            
                     // La URL debe cambiar
-                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/10/create?step=2`);
+                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/23/create?step=2`);
             
                     // El titulo de firmantes debe estar visible
                     await expect(page.locator('h1').filter({hasText: 'FIRMANTES'})).toBeVisible();
@@ -286,6 +270,7 @@ test.describe.serial('Certificados - Financieros Reinvertidas - Pruebas con los 
             
                     // Cerrar los mensajes que aparecen
                     await page.locator(`${ariaCerrar}`).first().click();
+                    await page.locator(`${ariaCerrar}`).last().click();
             
                     // Boton de Agregar Firmantes debe estar visible
                     const botonAgregarFirmantes = page.locator('text=Agregar Firmante');
@@ -360,7 +345,7 @@ test.describe.serial('Certificados - Financieros Reinvertidas - Pruebas con los 
             
                 test('Crear una Nueva Cuenta de Certificado - Paso 3 - Metodo de Interes', async () => {
                     // La URL debe cambiar
-                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/10/create?step=3`);
+                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/23/create?step=3`);
             
                     // El titulo principal debe estar visible
                     await expect(page.locator('h1').filter({hasText: 'FORMA PAGO DE INTERESES O EXCEDENTES'})).toBeVisible();
@@ -372,7 +357,7 @@ test.describe.serial('Certificados - Financieros Reinvertidas - Pruebas con los 
                 test('Finalizar con la Creacion de Cuenta de Certificado', async () => {
                     // Esperar que el mensaje de que los contratos se hayan generado se muestre
                     await expect(page.locator('text=Contratos Generados Exitosamente.')).toBeVisible();
-                    
+
                     // Boton de Finalizar
                     const botonFinalizar = page.locator('button:has-text("Finalizar")');
                     // Esperar que se abran dos pestañas
@@ -389,7 +374,7 @@ test.describe.serial('Certificados - Financieros Reinvertidas - Pruebas con los 
                     await newPage2.close();
 
                     // Debe regresar a la pagina de los certificados
-                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/10`);
+                    await expect(page).toHaveURL(`${url_base}/crear_cuentas/01-2-5-4/certificados/23`);
                 });
             };
